Web Version of Game
I originally intended for this game to be played as a windows .exe file and only worked on web compatibility after the game was already complete. This shows a bit since I used some rather basic methods to just get it working rather than focusing on making it more suited for the web.